+++ packages/rev-api/src/api/manager.ts
@@ -1,6 +1,6 @@
 
 import { IModelManager, IModel } from 'rev-models';
-import { checkIsModelConstructor } from 'rev-models/lib/models/utils';
+import { checkIsValidModelConstructor } from 'rev-models/lib/models/utils';
 import { initialiseApiMeta } from '../api/meta';
 import { IModelApiManager, IApiMeta } from './types';
 import { GraphQLApi } from '../graphql/api';
@@ -31,7 +31,7 @@ export class ModelApiManager implements IModelApiManager {
 
     register<T extends IModel>(model: new(...args: any[]) => T, apiMeta?: Partial<IApiMeta>) {
         // Add api meta to the registry if valid
-        checkIsModelConstructor(model);
+        checkIsValidModelConstructor(model);
         const meta = initialiseApiMeta(this, model, apiMeta);
         this._apiMeta[meta.model] = meta;
     }