+++ packages/rev-models/src/models/utils.ts
@@ -4,9 +4,19 @@ import { IModelMeta } from './types';
 /**
  * @private
  */
-export function checkIsModelConstructor(model: any) {
+export function checkIsValidModelConstructor(model: any) {
     if (!model || typeof model != 'function' || !model.name) {
         throw new Error('ModelError: Supplied model is not a model constructor.');
+    }
+    let instance: any;
+    try {
+        instance = new model();
+    }
+    catch (e) {
+        throw new Error(`ModelError: constructor threw an error when called with no arguments: ${e.message}`);
+    }
+    if (Object.keys(instance).length) {
+        throw new Error('ModelError: constructor must not set default field values. You should use a defaults() method instead.');
     }
 }