+++ packages/CHANGELOG.md
@@ -6,6 +6,7 @@
 **`rev-models`**
 
 * `ModelManager.isNew()` throw better error when model is null or undefined
+* `ModelManager.register()` now does a test instantiation of the model, and throws an error if the constructor sets any default values
 
 **`rev-ui`**